Задание 2. Изменены отступы на пробелы, все классы перенесены в пакеты и в классах добавлен автор.
