--- src/MyApplication.java
@@ -1,7 +1,0 @@
-import com.tinkoff.edu.Printer;
-
-public class MyApplication {
-	public static void main(String[] args) {
-		Printer.print("Hello, World!");
-	}
-}
