+++ src/com/tinkoff/edu/MyApplication.java
@@ -1,0 +1,14 @@
+package com.tinkoff.edu;
+
+import com.tinkoff.edu.helpers.Printer;
+
+/**
+ * Created on 08.08.2021
+ *
+ * @author Natalya Vinogradova
+ */
+public class MyApplication {
+    public static void main(String[] args) {
+        Printer.print("Hello, World!");
+    }
+}
