--- src/com/tinkoff/edu/Printer.java
@@ -1,7 +1,0 @@
-package com.tinkoff.edu;
-
-public class Printer {
-	public static void print(String message) {
-		System.out.println(message);
-	}
-}
