+++ src/com/tinkoff/edu/helpers/Printer.java
@@ -1,0 +1,12 @@
+package com.tinkoff.edu.helpers;
+
+/**
+ * Created on 08.08.2021
+ *
+ * @author Natalya Vinogradova
+ */
+public class Printer {
+    public static void print(String message) {
+        System.out.println(message);
+    }
+}
